Ranking base

+++ lib/bar_graph/individual_bar.dart
@@ -1,7 +1,7 @@
 
 class IndividualBar {
   final int x;
-  final double y;
+  final int y;
 
   IndividualBar({required this.x, required this.y});
 }
